feat(contract): implement reputation tests and LP token calculator
